+++ symfony/tests/Api/Console/Blog/Webhook/DeleteWebhookTest.php
@@ -1,0 +1,52 @@
+<?php
+
+namespace App\Tests\Api\Console\Blog\Webhook;
+
+use App\Api\Console\Controller\WebhookController;
+use App\Tests\Case\ApiTestCase;
+use App\Tests\Factory\BlogFactory;
+use App\Tests\Factory\WebhookFactory;
+use Hyvor\Internal\Auth\AuthFake;
+use PHPUnit\Framework\Attributes\CoversClass;
+
+#[CoversClass(WebhookController::class)]
+class DeleteWebhookTest extends ApiTestCase
+{
+    public function test_delete_webhook(): void
+    {
+        [$blog, $user] = BlogFactory::createOneWithUser(
+            ['subdomain' => 'wh-delete'],
+            ['hyvor_user_id' => 106, 'status' => 'active'],
+        );
+        $webhook = WebhookFactory::createOne([
+            'blog' => $blog,
+            'blog_id' => $blog->getId(),
+        ]);
+
+        $authUser = AuthFake::generateUser(['id' => 106]);
+        $this->consoleBlogApi('DELETE', 'wh-delete', '/webhook/' . $webhook->getId(), user: $authUser);
+
+        $this->assertResponseIsSuccessful();
+    }
+
+    public function test_delete_webhook_wrong_blog(): void
+    {
+        [$blog1, $user1] = BlogFactory::createOneWithUser(
+            ['subdomain' => 'wh-del-b1'],
+            ['hyvor_user_id' => 107, 'status' => 'active'],
+        );
+        [$blog2, $user2] = BlogFactory::createOneWithUser(
+            ['subdomain' => 'wh-del-b2'],
+            ['hyvor_user_id' => 108, 'status' => 'active'],
+        );
+        $webhook = WebhookFactory::createOne([
+            'blog' => $blog2,
+            'blog_id' => $blog2->getId(),
+        ]);
+
+        $authUser = AuthFake::generateUser(['id' => 107]);
+        $this->consoleBlogApi('DELETE', 'wh-del-b1', '/webhook/' . $webhook->getId(), user: $authUser);
+
+        $this->assertResponseStatusCodeSame(404);
+    }
+}
